feat: 48강까지 완료

+++ src/main/java/com/ll/sbbrestapi20250106/domain/question/QuestionController.java
@@ -1,17 +1,18 @@
 package com.ll.sbbrestapi20250106.domain.question;
 
-import com.ll.sbbrestapi20250106.domain.question.dto.QuestionDto;
+import com.ll.sbbrestapi20250106.domain.question.dto.QuestionDetailDto;
+import com.ll.sbbrestapi20250106.domain.question.dto.QuestionListDto;
 import com.ll.sbbrestapi20250106.domain.user.SiteUser;
 import com.ll.sbbrestapi20250106.domain.user.UserService;
+import com.ll.sbbrestapi20250106.global.rq.Rq;
 import com.ll.sbbrestapi20250106.global.rsData.RsData;
-import jakarta.transaction.Transactional;
+import com.ll.sbbrestapi20250106.standard.page.PageDto;
+import org.springframework.transaction.annotation.Transactional;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.validator.constraints.Length;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("/api/v1/question_list")
@@ -20,25 +21,45 @@ public class QuestionController {
 
     private final QuestionService questionService;
     private final UserService userService;
+    private final Rq rq;
 
     @GetMapping
-    public List<QuestionDto> getList() {
-        return questionService.findAll()
-                .stream()
-                .map(QuestionDto::new)
-                .toList();
+    @Transactional(readOnly = true)
+    public PageDto<QuestionListDto> getList(
+            @RequestParam(defaultValue = "1") int page,
+            @RequestParam(defaultValue = "10") int pageSize,
+            @RequestParam(defaultValue = "subject") String searchKeywordType,
+            @RequestParam(defaultValue = "") String searchKeyword
+    ) {
+        return new PageDto<>(
+                questionService.findByListedPaged(true, searchKeywordType, searchKeyword, page, pageSize)
+                        .map(QuestionListDto::new)
+        );
     }
 
     @GetMapping("/{id}")
-    public QuestionDto getDetail(@PathVariable Long id) {
-        return questionService.findById(id)
-                .map(QuestionDto::new)
-                .orElseThrow();
+    @Transactional(readOnly = true)
+    public QuestionDetailDto getDetail(@PathVariable Long id) {
+        Question question = questionService.findById(id).get();
+
+        if (!question.isPublished()) {
+            SiteUser user = rq.checkAuthentication();
+
+            question.checkActorCanRead(user);
+        }
+
+        return new QuestionDetailDto(question);
     }
 
     @DeleteMapping("/{id}")
+    @Transactional
     public RsData<Void> deleteQuestion(@PathVariable Long id) {
+        SiteUser user = rq.checkAuthentication();
+
         Question question = questionService.findById(id).get();
+
+        question.checkActorCanDelete(user);
+
         questionService.delete(question);
 
         return new RsData<>(
@@ -53,22 +74,30 @@ public class QuestionController {
             String subject,
             @NotBlank
             @Length(min = 3)
-            String content
+            String content,
+            boolean published,
+            boolean listed
     ) {}
 
 
     @PutMapping("/{id}")
     @Transactional
-    public RsData<QuestionDto> modifyQuestion(@PathVariable Long id,
-                                 @RequestBody @Valid QuestionModifyReqBody reqBody) {
+    public RsData<QuestionDetailDto> modifyQuestion(@PathVariable Long id,
+                                                  @RequestBody @Valid QuestionModifyReqBody reqBody) {
+        SiteUser user = rq.checkAuthentication();
+
         Question question = questionService.findById(id).get();
 
-        questionService.modify(question, reqBody.subject, reqBody.content);
+        question.checkActorCanModify(user);
+
+        questionService.modify(question, reqBody.subject, reqBody.content, reqBody.published, reqBody.listed);
+
+        questionService.flush();
 
         return new RsData<>(
                 "200-1",
                 "%d번 글이 수정되었습니다.".formatted(id),
-                new QuestionDto(question)
+                new QuestionDetailDto(question)
         );
     }
 
@@ -78,26 +107,22 @@ public class QuestionController {
             String subject,
             @NotBlank
             @Length(min = 3)
-            String content
-    ) {}
-
-    record QuestionCreateResBody (
-            QuestionDto item,
-            long totalCount
+            String content,
+            boolean published,
+            boolean listed
     ) {}
 
     @PostMapping
-    public RsData<QuestionCreateResBody> createQuestion(@RequestBody @Valid QuestionCreateReqBody reqBody) {
-        SiteUser user = userService.findByUsername("user3").get();
-        Question question = questionService.write(reqBody.subject, reqBody.content, user);
+    @Transactional
+    public RsData<QuestionDetailDto> createQuestion(@RequestBody @Valid QuestionCreateReqBody reqBody) {
+        SiteUser user = rq.checkAuthentication();
+
+        Question question = questionService.write(user, reqBody.subject, reqBody.content, reqBody.published, reqBody.listed);
 
         return new RsData<>(
                 "201-1",
                 "%d번 글이 작성되었습니다.".formatted(question.getId()),
-                new QuestionCreateResBody(
-                        new QuestionDto(question),
-                        questionService.count()
-                )
+                new QuestionDetailDto(question)
         );
     }
